Refactor imports and update dependencies
